Refactor server router module
